+++ R/C5.0.R
@@ -86,8 +86,21 @@ c5_fit  <- function(split, spec, control = control_bag()) {
   if (control$sampling == "down") {
     dat <- down_sampler(dat)
   }
+  if (any(names(dat) == ".weights")) {
+    wts <- hardhat::importance_weights(dat$.weights)
+    dat$.weights <- NULL
+  } else {
+    wts <- NULL
+  }
 
-  mod <- parsnip::fit.model_spec(spec, .outcome ~ ., data = dat, control = ctrl)
+  mod <-
+    parsnip::fit.model_spec(
+      spec,
+      .outcome ~ .,
+      data = dat,
+      control = ctrl,
+      case_weights = wts
+    )
   mod
 }