+++ R/bagger.R
@@ -12,6 +12,8 @@
 #'  underlying model function.
 #' @param data A data frame containing the variables used in the formula or
 #'  recipe.
+#' @param weights A numeric vector of non-negative case weights. These values are
+#' not used during bootstrap resampling.
 #' @param base_model A single character value for the model being bagged. Possible
 #'  values are "CART", "MARS", "C5.0" (classification only), or "nnet".
 #' @param times A single integer greater than 1 for the maximum number of bootstrap
@@ -110,8 +112,7 @@ bagger.default <- function(x, ...) {
 #' @export
 #' @rdname bagger
 bagger.data.frame <-
-  function(x,
-           y,
+  function(x, y, weights = NULL,
            base_model = "CART",
            times = 11L,
            control = control_bag(),
@@ -124,6 +125,7 @@ bagger.data.frame <-
     processed <- hardhat::mold(x, y)
     res <-
       bagger_bridge(processed,
+                    weights,
                     base_model,
                     seed,
                     times,
@@ -140,8 +142,7 @@ bagger.data.frame <-
 #' @export
 #' @rdname bagger
 bagger.matrix <-
-  function(x,
-           y,
+  function(x, y, weights = NULL,
            base_model = "CART",
            times = 11L,
            control = control_bag(),
@@ -155,6 +156,7 @@ bagger.matrix <-
     processed <- hardhat::mold(x, y)
     res <-
       bagger_bridge(processed,
+                    weights,
                     base_model,
                     seed,
                     times,
@@ -171,8 +173,7 @@ bagger.matrix <-
 #' @export
 #' @rdname bagger
 bagger.formula <-
-  function(formula,
-           data,
+  function(formula, data, weights = NULL,
            base_model = "CART",
            times = 11L,
            control = control_bag(),
@@ -187,6 +188,7 @@ bagger.formula <-
     processed <- hardhat::mold(formula, data, blueprint = bp)
     res <-
       bagger_bridge(processed,
+                    weights,
                     base_model,
                     seed,
                     times,
@@ -203,8 +205,7 @@ bagger.formula <-
 #' @export
 #' @rdname bagger
 bagger.recipe <-
-  function(x,
-           data,
+  function(x, data,
            base_model = "CART",
            times = 11L,
            control = control_bag(),
@@ -218,6 +219,7 @@ bagger.recipe <-
     processed <- hardhat::mold(x, data)
     res <-
       bagger_bridge(processed,
+                    weights = NULL,
                     base_model,
                     seed,
                     times,