+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Bagged single layer, feed-forward neural networks can be used with `base_model = "nnet"`.
 
+* Enable case weights for `bagger()`. 
+
 # baguette 0.2.0
 
 * Model definitions were moved to parsnip (#47)